api/requirements.txt
requests
paho-mqtt
json
flask